Trash `script/*` scripts in favor of Makefile
The current `script/bootstrap` and `script/build` are unfriendly to Go
newbies because they require:

- The user having a GOPATH
- The user having `godep`
- The user having `hg` to fetch godep

Ideally, the build process should work even if the user doesn't have any
of that. This changes `script/bootstrap` to simply check environment for
prerequisites (go 1.2, make), and relies on `make hub` for the build
process.